% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRIP.catch.R
\name{calculate_catch}
\alias{calculate_catch}
\title{Calculate catch}
\usage{
calculate_catch(
  intdir = NULL,
  common = NULL,
  st = NULL,
  styr = NULL,
  endyr = NULL,
  dom = NULL
)
}
\arguments{
\item{intdir}{data directory}

\item{common}{species names(s)}

\item{st}{The state(s)}

\item{styr}{Start year}

\item{endyr}{End year}

\item{dom}{optional custom domain}
}
\value{
a dataframe with catch estimates and PSEs
}
\description{
Calculate catch
}
