% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRIP.dirtrips.R
\name{MRIP.dirtrips}
\alias{MRIP.dirtrips}
\title{Calculate directed trips}
\usage{
MRIP.dirtrips(
  intdir = NULL,
  common = NULL,
  st = NULL,
  styr = NULL,
  endyr = NULL,
  trips = c(1, 2, 3, 4, 5),
  dom = NULL
)
}
\arguments{
\item{intdir}{data directory}

\item{common}{species names(s)}

\item{st}{The state(s)}

\item{styr}{Start year}

\item{endyr}{End year}

\item{trips}{type of trips}

\item{dom}{optional custom domain}
}
\value{
a dataframe with directed trip estimates and PSEs
}
\description{
Calculate directed trips
}
